Could not create new droplets #107
New Droplets are generated in  add_droplet.php.

In tools.php  line 29 they are deleted again, always ...
So the query in line 131 cannot return any values for this droplet as
its already deleted .
So the hidden field cannot be populated and no droplet_id is sent to save.

So the Droplet is created , deleted , and then tried to edit ...

I added an if clause that only deletes the empty droplets on normal tool
view, whithout do=something. This should do it.  Hopefully.

+++ wbce/modules/droplets/tool.php
@@ -26,7 +26,10 @@ require_once dirname(__FILE__).'/functions.inc.php';
 
 // removes empty entries from the table; this may happen if someone adds a
 // droplet but does not save any changes
-$database->query(sprintf("DELETE FROM `%smod_droplets` WHERE name=''",TABLE_PREFIX));
+if (!$admin->get_get("do")) {
+    $database->query(sprintf("DELETE FROM `%smod_droplets` WHERE name=''",TABLE_PREFIX));
+}
+
 
 $twig_data = array(
     'FTAN' => (method_exists($admin,'getFTAN') ? $admin->getFTAN() : '')
